feat: #186 - 접근 제어 유틸 개선 및 getUser 기반 인증 검증 적용
