+++ src/app/(auth)/reset-password/page.tsx
@@ -3,7 +3,7 @@ import { Suspense } from "react";
 
 import { resetPasswordAction } from "@/features/auth/reset-password/actions/resetPasswordAction";
 import { ResetPasswordForm } from "@/features/auth/reset-password/components/ResetPasswordForm";
-import { requireAuthSession } from "@/lib/utils/requireAuthSession";
+import { requireAuthUser } from "@/features/auth/utils/requireAuthUser";
 
 // 검색 엔진 인덱싱 방지 (robots.txt Disallow보다 확실함 — 삭제 금지)
 export const metadata: Metadata = {
@@ -39,8 +39,9 @@ type Props = {
  * 오류가 발생한다.
  */
 export default async function ResetPasswordPage({ searchParams }: Props) {
-  // reset-password 직접 접근 시 session이 없으면 login으로 redirect
-  await requireAuthSession();
+  // 인증되지 않은 사용자는 login으로 redirect
+  await requireAuthUser();
+
   /**
    * redirect query 전달
    *
